X: Undo webhooks.events schema change

+++ database/migrations/2015_11_24_225631_create_webhooks.php
@@ -33,7 +33,7 @@ class CreateWebhooks extends Migration
 
             $table->string(Webhook::URL);
 
-            $table->bigInteger(Webhook::EVENTS);
+            $table->integer(Webhook::EVENTS);
 
             $table->string(Webhook::ENTITY_TYPE, 100)
                   ->nullable();
